ASoC: SOF: clients: Add API to manage the module refcount of SOF core
The SOF core module (the device driver of the DSP) needs to be protected
against removal when a client is active and needs the core to be present
during it's active operation.

sof_client_core_moduleget/put can be use by client drivers to manage the
refcount on the core's module to prevent such a case.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>