ASoC: SOF: Add library file to handle the firmware file profiles
The purpose of the library is to craft the paths and firmware file name
for the IPC version.
After creating the paths we will try to open the default firmware image
ad check it's magic number to make sure it is matching with the expected
IPC type.

The sof_create_default_fw_layout_profile() will try to first create the
strings for the requested IPC type.
If no layout profile was found for the IPC type then we will try to probe
other IPC types as a fallback.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>