ASoC: SOF: ipc4-topology: Convert FLOAT to S32 during blob selection
SSP/DMIC blobs have no support for FLOAT type, they are using S32 on data
bus.

Convert the format from FLOAT_LE to S32_LE to make sure that the correct
format is used within the path.

FLOAT conversion will be done on the host side (or within the path).

Fixes: f7c41911ad74 ("ASoC: SOF: ipc4-topology: Add support for float sample type")
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>