ASoC: SOF: Intel: hda: Only check SSP MCLK mask in case of IPC3
IPC4 is using the NHLT blob itself and sends the SSP blob from it directly
to the firmware, there is no need for the MCLK quirk based on the SSP blob
since the SSP blob is in use.

At the same time reword the error, info and debug messages for clarity.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>