ASoC: SOF: ipc4-compress: Complete the pointer support
The pointer support also needs to report the PCM progression on the DAI
side.
We can re-use the code from ipc4-pcm delay reporting with some modification
for the compr stream to get the DAI frame counter.
In case of compr there is no definition of delay, so the code can be a bit
simpler as we need to care about the two pointer separately.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>